Resetting MainConfig setters after using .setMainConfig(config) should not reset to defaults.
With a usage pattern of new AutomationRunner(data).setMainConfig(modelConfig).setParallelism(18)
prior to this change, all defaults would be used if a setter were used after .setMainConfig() was applied.
This change mutates all class level var's to the values specified in the .setMainConfig() configuration supplied.
This will allow the use of setters to modify the object's parameters without reverting anything to defaults.