Update FeatureCorrelationDetection.scala
In the previous code, both of the features having higher negative or positive correlation were getting removed because for every pair of feature a and b, in correlationInteraction (a,b) as well as (b,a) was getting added. But we want to remove only one of the two features either a or b.